feat: integration

+++ sipel/src/main/java/propen/impl/sipel/restcontroller/ReportRestController.java
@@ -1,6 +1,5 @@
 package propen.impl.sipel.restcontroller;
 
-import org.apache.tomcat.util.http.fileupload.FileUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
@@ -20,11 +19,10 @@ import propen.impl.sipel.service.ReportRestService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
-import java.io.*;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.nio.channels.Channels;
-import java.nio.channels.ReadableByteChannel;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -32,8 +30,11 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.logging.Logger;
 
+import org.springframework.security.access.prepost.PreAuthorize;
+
+@CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
-@CrossOrigin(origins = "*")
+//@RequestMapping("")
 public class ReportRestController {
 
     @Autowired
@@ -42,9 +43,12 @@ public class ReportRestController {
     @Autowired
     FileStorageService fileStorageService;
 
+    private static final Logger logger = Logger.getLogger(ReportRestController.class.getName());
+
     // Mengembalikan list report yang berjenis installation dan maintenance
     @GetMapping(value="/api/v1/reportsIrMr")
-    private List<ReportModel> retrieveListReportIrMr(){
+    @PreAuthorize("hasRole('ENGINEER')")
+    public List<ReportModel> retrieveListReportIrMr(){
         List<ReportModel> listReport = reportRestService.retrieveListReport();
 
         List<ReportModel> listReportFiltered = new ArrayList<>();
@@ -55,14 +59,15 @@ public class ReportRestController {
             }
         }
 
-        return listReport;
+        return listReportFiltered;
     }
 
     // Menyimpan file yang diupload ke local server dan membuat report baru
     // File yang memiliki nama yang sama akan dibuat nama dengan versi
     // Mengembalikan response dengan result report yang berhasil dibuat
     @PostMapping(value="/api/v1/report/upload", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    private BaseResponse<ReportModel> uploadReport(@Valid @ModelAttribute ReportDto report,
+    @PreAuthorize("hasRole('ENGINEER')")
+    public BaseResponse<ReportModel> uploadReport(@Valid @ModelAttribute ReportDto report,
                                                    HttpServletRequest request) throws Exception{
         BaseResponse<ReportModel> response = new BaseResponse<>();
         if(report.getReportType() == null && report.getFile() == null){
@@ -74,7 +79,6 @@ public class ReportRestController {
 
         // Root Directory
         String uploadRootPath = request.getServletContext().getRealPath("upload");
-//        System.out.println("uploadRootPath=" + uploadRootPath);
 
         File uploadRootDir = new File(uploadRootPath);
         // Create directory if it not exists.
@@ -95,13 +99,9 @@ public class ReportRestController {
                 fileNameOriginal = listFileNameOriginal[0] + " ver.2" + "." + listFileNameOriginal[1];
             }
         }
-//        String fileName = fileStorageService.storeFile(report.getFile(), fileNameOriginal);
-        String fileName = fileStorageService.storeFile(uploadRootDir, fileNameOriginal, report.getFile());
-        String urlFile = ServletUriComponentsBuilder.fromCurrentContextPath()
-                .path("/report/")
-                .path(fileName)
-                .toUriString();
-        report.setReportName(fileName);
+        File file = fileStorageService.storeFile(uploadRootDir, fileNameOriginal, report.getFile());
+        String urlFile = file.getAbsolutePath();
+        report.setReportName(fileNameOriginal);
         report.setFileType(report.getFile().getContentType());
         report.setSize(report.getFile().getSize());
         ReportModel newReport = reportRestService.uploadReport(report, urlFile);
@@ -115,39 +115,8 @@ public class ReportRestController {
     // Download file report yang dipilih
     @GetMapping("/report/{fileName:.+}")
     public ResponseEntity<Resource> downloadReport(@PathVariable String fileName) throws IOException {
-//        Resource resource = fileStorageService.loadFileAsResource(fileName);
-//        ReportModel report = reportRestService.findReportByReportName(fileName);
-//        Resource resource = fileStorageService.loadFileAsResource(report.getUrlFile(), fileName);
-//        String fileType = null;
-//
-//        try{
-//            fileType = request.getServletContext().getMimeType(resource.getFile().getAbsolutePath());
-//        }catch (IOException e){
-//            System.out.println("Tidak dapat menentukan tipe file");
-//        }
-//
-//        if(fileType==null){
-//            fileType = "application/octet-stream";
-//        }
 
         ReportModel reportTarget = reportRestService.findReportByReportName(fileName);
-//        BufferedInputStream in = null;
-//        FileOutputStream fout = null;
-//        try {
-//            in = new BufferedInputStream(new URL(reportTarget.getUrlFile()).openStream());
-//            fout = new FileOutputStream(fileName);
-//
-//            byte data[] = new byte[1024];
-//            int count;
-//            while ((count = in.read(data, 0, 1024)) != -1) {
-//                fout.write(data, 0, count);
-//            }
-//        } finally {
-//            if (in != null)
-//                in.close();
-//            if (fout != null)
-//                fout.close();
-//        }
 
         File file = new File(reportTarget.getUrlFile());
 
@@ -162,19 +131,13 @@ public class ReportRestController {
                 .contentLength(file.length())
                 .contentType(MediaType.APPLICATION_OCTET_STREAM)
                 .body(resource);
-//                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + fileName + "\"")
-//                .contentType(MediaType.parseMediaType(fileType))
-//                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + resource.getFilename() + "\"")
-//                .body(resource);
     }
 
     // Menampilkan preview dari file yang dipilih dan berjenis pdf tanpa men-download
     @GetMapping("/report/{fileName:.+}/preview")
     public ResponseEntity<InputStreamResource> previewReport(@PathVariable String fileName) throws FileNotFoundException {
-//        Path filePath = fileStorageService.getFilePath(fileName);
         ReportModel report = reportRestService.findReportByReportName(fileName);
-//        File file = new File(""+filePath+"");
-        File file = new File(""+report.getUrlFile()+"");
+        File file = new File(report.getUrlFile());
         HttpHeaders headers = new HttpHeaders();
         headers.add("content-disposition", "inline;filename=" +fileName);
 
@@ -189,22 +152,45 @@ public class ReportRestController {
 
     // Menghapus file dari local server dan report dari database
     @DeleteMapping(value="/api/v1/report/{idReport}/delete")
-    private ResponseEntity<String> deleteReport(@PathVariable("idReport") Long idReport) {
+    @PreAuthorize("hasRole('ENGINEER')")
+    public ResponseEntity<String> deleteReport(@PathVariable("idReport") Long idReport) {
         try{
             ReportModel report = reportRestService.findReportById(idReport);
-            String fileName = report.getReportName();
-            Path filePath = fileStorageService.getFilePath(fileName);
-            reportRestService.deleteReport(idReport);
-            Files.delete(filePath);
-            return ResponseEntity.ok("Report dengan ID "+String.valueOf(idReport)+" berhasil dihapus!");
-        }catch (NoSuchElementException | IOException e){
+            File file = new File(report.getUrlFile());
+            if(file.delete()){
+                reportRestService.deleteReport(idReport);
+                return ResponseEntity.ok("Report dengan ID "+String.valueOf(idReport)+" berhasil dihapus!");
+            }else{
+                throw new ResponseStatusException(
+                        HttpStatus.NOT_FOUND, "Report dengan ID "+String.valueOf(idReport)+" tidak ditemukan!"
+                );
+            }
+        }catch (NoSuchElementException e){
             throw new ResponseStatusException(
                     HttpStatus.NOT_FOUND, "Report dengan ID "+String.valueOf(idReport)+" tidak ditemukan!"
             );
         }
     }
+
+    @PutMapping(value = "/api/v1/report/update/{idReport}")
+    @PreAuthorize("hasRole('MANAGER')")
+    public ReportModel updateStatusReport(
+            @PathVariable (value = "idReport") Long idReport,
+            @RequestBody ReportModel report
+    ) {
+        try {
+            return reportRestService.updateStatus(idReport, report);
+        }
+        catch (NoSuchElementException e) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Report with ID " + String.valueOf(idReport) + " not found!"
+            );
+        }
+    }
+
     @GetMapping(value="/api/v1/reports")
-    private List<ReportModel> retrieveListReport(){
+    @PreAuthorize("hasRole('FINANCE')")
+    public List<ReportModel> retrieveListReportApproved(){
         List<ReportModel> listReport = reportRestService.retrieveListReport();
         List<ReportModel> toBeSeenReport = new ArrayList<>();
         for(ReportModel report : listReport){
@@ -216,8 +202,10 @@ public class ReportRestController {
 
         return toBeSeenReport;
     }
+
     @GetMapping(value="/api/v1/reports/all")
-    private List<ReportModel> retrieveListReportAll(){
+    @PreAuthorize("hasRole('MANAGER')")
+    public List<ReportModel> retrieveListReportAll(){
         List<ReportModel> listReport = reportRestService.retrieveListReport();
 
         return listReport;
